+++ SerialDeviceSwift/SerialDeviceSwift/ViewController.swift
@@ -1,0 +1,27 @@
+//
+//  ViewController.swift
+//  SerialDeviceSwift
+//
+//  Created by Artem Hruzd on 3/9/18.
+//  Copyright © 2018 Artem Hruzd. All rights reserved.
+//
+
+import Cocoa
+
+class ViewController: NSViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override var representedObject: Any? {
+        didSet {
+        // Update the view, if already loaded.
+        }
+    }
+
+
+}
+